pyproject.toml
[project]
name = "pairs-trading"
version = "0.1.0"
description = "Pairs Trading Algo using yfinance"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance (>=0.2.62,<0.3.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
black = "^25.1.0"

